serial: moving drivers to the right place
